Add performance bucket support for statistics views, menu navigation, and advanced stats filtering.
- Implement per-category filtering and grouping for stats.
- Add bucket selection UI in `StatsOverview` and `AdvancedStats`.
- Enhance server and bucket metadata handling across components.
- Extend data models and service queries to support bucket codes.
- Refactor side context menus with section headers for buckets.
- Optimize performance bucket normalization and validation logic.

+++ Plugins/ZombieStats/Events/ZombieEventProcessor.cs
@@ -295,9 +295,9 @@ public class ZombieEventProcessor(ILogger<ZombieEventProcessor> logger, ZombieCl
             // undo that
             if (gameEvent.IsGameOver && !isForfeit)
             {
-                var previousCumulativePoints = (int)matchStat.PointsEarned;
+                var previousCumulativePoints = matchStat.PointsEarned;
                 var lastRoundPoints = gameEvent.TotalScore - previousCumulativePoints;
-                currentScore = lastRoundPoints + (previousCumulativePoints - (int)matchStat.PointsSpent);
+                currentScore = (int)(lastRoundPoints + (previousCumulativePoints - matchStat.PointsSpent));
             }
 
             roundState.PersistentClientRound.Points = currentScore;
